Simplify contracts-interface and refactor examples
