+++ .eslintrc
@@ -32,7 +32,8 @@
 		"semi": ["error", "always"],
 		"no-console": "error",
 		"@typescript-eslint/ban-ts-comment": "off",
-		"@typescript-eslint/explicit-module-boundary-types": "off"
+		"@typescript-eslint/explicit-module-boundary-types": "off",
+		"@typescript-eslint/no-var-requires": "off"
 	},
 	"settings": {
 		"import/resolver": {
