+++ package.json
@@ -66,10 +66,5 @@
 			"eslint"
 		]
 	},
-	"optionalDependencies": {
-		"chokidar": "^3.5.2",
-		"fsevents": "^2.3.2",
-		"watchpack-chokidar2": "^2.0.1"
-	},
 	"packageManager": "yarn@3.2.1"
 }
