--- packages/contracts-interface/examples/.eslintrc
@@ -1,6 +1,0 @@
-{
-	"rules": {
-		"no-console": "off",
-		"@typescript-eslint/no-var-requires": "off"
-	}
-}
