--- packages/contracts-interface/tsconfig.node.json
@@ -1,10 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"declaration": true,
-		"declarationMap": true,
-		"module": "commonjs",
-		"outDir": "./build/node/src"
-	},
-	"include": ["src"]
-}
