+++ tools/deps/.eslintrc
@@ -1,6 +1,5 @@
 {
 	"rules": {
-		"no-console": "off",
-		"@typescript-eslint/no-var-requires": "off"
+		"no-console": "off"
 	}
 }
