--- tools/safe-import/.eslintrc
@@ -1,5 +1,0 @@
-{
-	"rules": {
-		"@typescript-eslint/no-var-requires": "off"
-	}
-}
